Shift GUI-related classes to separate package
